padding: top right bottom left,padding-left:leftpx
<!-- padding: it is very similar to margin, instead of moving the entire
div it moves the text inside the div. Also if we look at the page we
know that there is too much distance in the text inside the div tag that
is because of p tag , it has padding inbuilt so we would have to set the
padding and margin of the p tag to be zero. you can do padding : top
right bottom left , and if you only give two out of 4 paddings it will
repeat the padding  . you can also do padding-left: leftpx;
padding-right: rightpx and etc-->